pyproject.toml
[project]
name = "goshdarnhomedirectoryisamessagain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.12"
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.46.1",
    "clean-text>=0.7.1",
    "docx-python>=0.1.0",
    "emoji>=2.15.0",
    "filetype>=1.2.0",
    "keybert>=0.9.0",
    "keyphrase-vectorizers>=0.0.9",
    "llama-cpp-python[server]>=0.3.16",
    "mammoth>=1.11.0",
    "numpy>=2.4.2",
    "python-docx>=1.2.0",
    "qwen-vl-utils>=0.0.14",
    "regex>=2026.1.15",
    "spacy>=3.8.11",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "transformers>=5.1.0",
]
